cloud/requirements.txt
fastapi>=0.110
uvicorn>=0.27
graphiti-core>=0.5
neo4j>=5.17
qdrant-client>=1.8
google-cloud-pubsub>=2.21
google-cloud-bigquery>=3.19
google-cloud-storage>=2.16
google-cloud-aiplatform>=1.43
langchain>=0.2
langgraph>=0.1
langchain-google-genai>=1.0
numpy>=1.26
scipy>=1.12
scikit-learn>=1.4
torch>=2.2
transformers>=4.38
open-clip-torch>=2.24
pydantic>=2.6
httpx>=0.27
PyJWT>=2.8
sentence-transformers>=2.5
